pyproject.toml
[tool.poetry]
name = "ai-agent-platform"
version = "0.1.0"
description = "Multi-provider AI Agent Platform"
authors = ["Daniel Melo <dreimao4@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
langchain = "^0.3.0"
langchain-anthropic = "^0.3.0"
langchain-google-genai = "^2.0.0"
qdrant-client = "^1.12.0"
python-dotenv = "^1.0.0"
pydantic = "^2.10.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
